# Conflicts:
#	cmake/ospray.cmake
#	ospray/CMakeLists.txt
#	ospray/common/OSPConfig.h.in
#	ospray/embree/common/simd/sse.h
#	ospray/mpi/MPILoadBalancer.cpp
#	ospray/mpi/worker.cpp
#	ospray/render/LoadBalancer.cpp
#	ospray/render/LoadBalancer.h
#	ospray/render/Renderer.cpp
#	ospray/render/Renderer.h
#	ospray/render/Renderer.ispc

+++ ospray/common/Core.cpp
@@ -50,7 +50,7 @@ namespace ospray {
       the master) that only coordinate rendering but don't hold data
       of their own. In localdevice mode, this will return '1'.
      */
-     size_t getWorkerCount()
+     int64 getWorkerCount()
      { 
 #ifdef OSPRAY_MPI
        return mpi::worker.size;
